+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fix "change committed" link in the latest activities ([#21](https://github.com/scm-manager/scm-landingpage-plugin/pull/21))
+
 ## 1.3.0 - 2020-12-04
 ### Added
 - Add repository import event for successful and failed imports ([#19](https://github.com/scm-manager/scm-landingpage-plugin/pull/19))
@@ -13,13 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show landingpage for anonymous user ([#18](https://github.com/scm-manager/scm-landingpage-plugin/pull/18))
 
 ## 1.1.0 - 2020-07-03
-
 ### Added
 - Add repository renamed event [#14](https://github.com/scm-manager/scm-landingpage-plugin/pull/14)
 
 ### Fixed
 - Add permission check before accessing repository data ([#15](https://github.com/scm-manager/scm-landingpage-plugin/pull/15))
-
 
 ## 1.0.0 - 2020-06-04
 ### Fixed
@@ -35,7 +37,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deleted repositories are now also removed from all users' favourites. ([#12](https://github.com/scm-manager/scm-landingpage-plugin/pull/12))
 
 ## 1.0.0-rc1 - 2020-04-14
-
 Initial release
 
 [1.0.0-rc1]: https://github.com/scm-manager/scm-landingpage-plugin/releases/tag/1.0.0-rc1